refactor(cli): restructure as standalone UV project with modern src/ layout
