+++ .github/workflows/publish.yml
@@ -1,9 +1,6 @@
 name: Publish to PyPI
 
 on:
-  push:
-    tags:
-      - "v*"
   workflow_dispatch:
 
 permissions:
@@ -14,21 +11,25 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v4
-      - name: Set up Python
-        uses: actions/setup-python@v5
+      - name: Install uv
+        uses: astral-sh/setup-uv@v3
         with:
-          python-version: "3.11"
+          version: "latest"
+      
+      - name: Set up Python
+        run: uv python install 3.13
 
-      - name: Install dependencies
+      - name: Build package
         run: |
-          python -m pip install --upgrade pip
-          pip install build twine
+          cd cli
+          uv build
 
-      - name: Build and publish
+      - name: Publish to PyPI
         env:
           TWINE_USERNAME: "__token__"
           TWINE_PASSWORD: ${{ secrets.PYPI_API_TOKEN }}
         run: |
-          python -m build
-          twine check dist/*
-          twine upload dist/*
+          cd cli
+          uv tool install twine
+          uv tool run twine check dist/*
+          uv tool run twine upload dist/*
