+++ .github/workflows/test.yml
@@ -1,10 +1,12 @@
-name: Tests
+name: CLI Tests
 
 on:
   push:
     branches: [master]
+    paths: ["cli/**"]
   pull_request:
     branches: [master]
+    paths: ["cli/**"]
 
 permissions:
   contents: read
@@ -14,26 +16,42 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        python-version: ["3.9", "3.13"]
+        python-version: ["3.13"]
 
     steps:
       - uses: actions/checkout@v4
-      - name: Set up Python ${{ matrix.python-version }}
-        uses: actions/setup-python@v5
+      
+      - name: Install uv
+        uses: astral-sh/setup-uv@v3
         with:
-          python-version: ${{ matrix.python-version }}
+          version: "latest"
+      
+      - name: Set up Python ${{ matrix.python-version }}
+        run: uv python install ${{ matrix.python-version }}
 
-      - name: Install dependencies
+      - name: Install CLI package
         run: |
-          python -m pip install --upgrade pip
-          pip install build twine
-          pip install -e .
+          cd cli
+          uv sync
+          uv pip install -e .
 
-      - name: Verify installation
+      - name: Verify CLI installation
         run: |
-          infragpt --help
+          cd cli
+          uv run infragpt --help
 
-      - name: Check package structure
+      - name: Run tests (if they exist)
         run: |
-          python -m build --sdist
-          twine check dist/*
+          cd cli
+          if [ -d "tests" ]; then
+            uv run pytest tests/ -v
+          else
+            echo "No tests directory found, skipping tests"
+          fi
+
+      - name: Check package build
+        run: |
+          cd cli
+          uv build
+          uv tool install twine
+          uv tool run twine check dist/*
