+++ cli/publish.sh
