+++ cli/src/infragpt/__main__.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 """Command-line entry point for InfraGPT."""
 
-from cli.cli import cli
+from infragpt.main import cli
 
 if __name__ == "__main__":
     cli()
