+++ cli/src/infragpt/bump_version.py
