+++ cli/src/infragpt/history.py
