+++ cli/src/infragpt/llm/__init__.py
@@ -10,23 +10,23 @@ and prompt handling.
 __version__ = "0.1.0"
 
 # Public exports from client.py
-from llm.client import (
+from infragpt.llm.client import (
     get_llm_client,
     generate_gcloud_command,
     get_parameter_info,
 )
 
 # Public exports from auth.py
-from llm.auth import validate_api_key
+from infragpt.llm.auth import validate_api_key
 
 # Public exports from prompts.py
-from llm.prompts import get_prompt_template, format_prompt
+from infragpt.llm.prompts import get_prompt_template, format_prompt
 
 # Public exports from models.py
-from llm.models import MODEL_TYPE
+from infragpt.llm.models import MODEL_TYPE
 
 # Public exports from errors.py
-from llm.errors import (
+from infragpt.llm.errors import (
     LLMError,
     AuthenticationError,
     GenerationError,
