+++ cli/src/infragpt/llm/errors.py
