--- requirements.txt
@@ -1,8 +1,0 @@
-langchain-core>=0.1.18
-langchain-openai>=0.0.5
-langchain-anthropic>=0.1.1
-rich>=13.7.0
-prompt-toolkit>=3.0.43
-click>=8.1.7
-pyperclip>=1.8.2
-pyyaml>=6.0
